QuietTube-1.1.0-local logging fix and improvment

+++ tests/test_settings_native.m
@@ -48,7 +48,7 @@ int main(int argc, const char *argv[]) {
             [writes removeAllObjects];
             NSDictionary *preset=QTPresetChanges(name);
             assert(writes.count==0);
-            assert(preset.count==([name isEqualToString:@"Focused feed"]?15:8));
+            assert(preset.count==(NSUInteger)([name isEqualToString:@"Focused feed"]?15:8));
             assert(!preset[@"background"] && !preset[@"autoplay"]);
             assert(![preset[@"mutationTrace"] boolValue] && ![preset[@"inspectElements"] boolValue]);
             for (NSString *key in @[@"enabled",@"adTest",@"feedAds",@"displayAds",@"extendedFeed",@"plainLogo"]) assert([preset[key] boolValue]);
